requirements.txt
psycopg2-binary
python-dotenv
loguru
slackclient
pytz
requests
twilio
wheel